+++ src/constants.ts
@@ -1,17 +1,8 @@
 export const LOCAL_WORKFLOWS_PATH = "./.github/workflows";
 export const UTF8 = "utf-8";
 
-export const WORKFLOW_NAMES = [
-    "on_demand_create_release_draft",
-    "on_pull_merged",
-    "on_pull_opened_or_edited",
-    "on_push_check",
-    "on_release_published",
-    "on_release_pull_merged"
-];
-export const HELP_WORKFLOW_NAMES = [
-    "all (all workflows below)",
-    ...WORKFLOW_NAMES
-];
+export const DOCS_URL = "https://github.com/vemel/github_actions_js";
 export const REPO_URL =
     "https://raw.githubusercontent.com/vemel/github_actions_js";
+export const JS_INDEX_URL = `${REPO_URL}/{ref}/workflows/index.yml`;
+export const PY_INDEX_URL = `${REPO_URL}/{ref}/workflows_py/index.yml`;