+++ src/utils.ts
@@ -1,4 +1,14 @@
+import path from "path";
+
 export function decapitalize(s: string): string {
     if (!s) return s;
     return `${s[0].toLowerCase()}${s.substr(1)}`;
+}
+
+export function joinURL(base: string, newPath: string): string {
+    const url = new URL(base);
+    let oldPathname = url.pathname;
+    if (!base.endsWith("/")) oldPathname = path.dirname(oldPathname);
+    const pathname = path.join(oldPathname, newPath);
+    return new URL(pathname, url.origin).href;
 }